Optimize task_comm and add gcs handler
